# Faster crc status