--- a/pkg/crc/cluster/clusteroperator_test.go
+++ b/pkg/crc/cluster/clusteroperator_test.go
@@ -1,12 +1,15 @@
 package cluster
 
 import (
+	"context"
+	"encoding/json"
 	"io/ioutil"
 	"path/filepath"
 	"testing"
 
-	"github.com/code-ready/crc/pkg/crc/oc"
+	v1 "github.com/openshift/api/config/v1"
 	"github.com/stretchr/testify/assert"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 var (
@@ -21,42 +24,32 @@ var (
 )
 
 func TestGetClusterOperatorsStatus(t *testing.T) {
-	status, err := GetClusterOperatorsStatus(ocConfig("co.json"))
+	status, err := getStatus(context.Background(), lister("co.json"), []string{})
 	assert.NoError(t, err)
 	assert.Equal(t, available, status)
 }
 
 func TestGetClusterOperatorsStatusProgressing(t *testing.T) {
-	status, err := GetClusterOperatorsStatus(ocConfig("co-progressing.json"))
+	status, err := getStatus(context.Background(), lister("co-progressing.json"), []string{})
 	assert.NoError(t, err)
 	assert.Equal(t, progressing, status)
 }
 
-func ocConfig(s string) oc.Config {
-	return oc.Config{
-		Runner: &mockRunner{file: filepath.Join("testdata", s)},
-	}
-}
-
-type mockRunner struct {
+type mockLister struct {
 	file string
 }
 
-func (r *mockRunner) Run(executablePath string, args ...string) (string, string, error) {
+func (r *mockLister) List(ctx context.Context, opts metav1.ListOptions) (*v1.ClusterOperatorList, error) {
 	bin, err := ioutil.ReadFile(r.file)
-	return string(bin), "", err
+	if err != nil {
+		return nil, err
+	}
+	var list v1.ClusterOperatorList
+	return &list, json.Unmarshal(bin, &list)
 }
 
-func (r *mockRunner) RunPrivate(executablePath string, args ...string) (string, string, error) {
-	bin, err := ioutil.ReadFile(r.file)
-	return string(bin), "", err
-}
-
-func (r *mockRunner) RunPrivileged(reason string, args ...string) (string, string, error) {
-	bin, err := ioutil.ReadFile(r.file)
-	return string(bin), "", err
-}
-
-func (r *mockRunner) GetKubeconfigPath() string {
-	return ""
+func lister(s string) *mockLister {
+	return &mockLister{
+		file: filepath.Join("testdata", s),
+	}
 }