… go.sum
… vendor/github.com/golang/protobuf/ptypes/any.go
… vendor/github.com/golang/protobuf/ptypes/any/any.pb.go
… vendor/github.com/golang/protobuf/ptypes/doc.go
… vendor/github.com/golang/protobuf/ptypes/duration.go
… vendor/github.com/golang/protobuf/ptypes/duration/duration.pb.go
… vendor/github.com/golang/protobuf/ptypes/timestamp.go
… vendor/github.com/golang/protobuf/ptypes/timestamp/timestamp.pb.go
… vendor/github.com/googleapis/gnostic/LICENSE
… vendor/github.com/googleapis/gnostic/compiler/README.md
… vendor/github.com/googleapis/gnostic/compiler/context.go
… vendor/github.com/googleapis/gnostic/compiler/error.go
… vendor/github.com/googleapis/gnostic/compiler/extension-handler.go
… vendor/github.com/googleapis/gnostic/compiler/helpers.go
… vendor/github.com/googleapis/gnostic/compiler/main.go
… vendor/github.com/googleapis/gnostic/compiler/reader.go
… vendor/github.com/googleapis/gnostic/extensions/README.md
… vendor/github.com/googleapis/gnostic/extensions/extension.pb.go
… vendor/github.com/googleapis/gnostic/extensions/extension.proto
… vendor/github.com/googleapis/gnostic/extensions/extensions.go
… vendor/github.com/googleapis/gnostic/openapiv2/OpenAPIv2.go
… vendor/github.com/googleapis/gnostic/openapiv2/OpenAPIv2.pb.go
… vendor/github.com/googleapis/gnostic/openapiv2/OpenAPIv2.proto
… vendor/github.com/googleapis/gnostic/openapiv2/README.md
… vendor/github.com/googleapis/gnostic/openapiv2/openapi-2.0.json
… vendor/github.com/kofalt/go-memoize/.gitignore
… vendor/github.com/kofalt/go-memoize/LICENSE.txt
… vendor/github.com/kofalt/go-memoize/go.mod
… vendor/github.com/kofalt/go-memoize/go.sum
… vendor/github.com/kofalt/go-memoize/memoize.go