`crc status` takes around 400-600ms on my machine. The macOS tray polls it every 5s. 
This PR makes crc status faster. 

It caches `df -h` output. The daemon no longer opens a ssh connection to get these numbers every 5s. The cache lasts 1 minute.
For the openshift status, the json is retrieved directly via the API server without using oc in ssh. The connection is setup to avoid proxies and use directly the VM IP.

For the CLI, crc status takes now around 200-300ms. From the daemon, 200ms without size cache, 50ms with cache.

`crc status` still hang from time to time (the bug was present before this PR - it happens mostly during crc start or stop when the connection is not yet ready). This is due to the ssh client not correctly handling timeouts. I will fix it in an other PR.