@cfergeau 2021-05-19
Just some high-level questions before I look more closely (not asking for changes)

> It caches df -h output. The daemon no longer opens a ssh connection to get these numbers every 5s. The cache lasts 1 minute.

An alternative could be for the daemon to keep a long lasting ssh connection, so that we don't need to recreate it for every status call? Or would it be too slow anyway?

> The macOS tray polls it every 5s. This PR makes crc status faster.

But does it need the full status all the time? I don't think disk size is required for each of these polls, so there could be a 'fast status' command and a slower one?