@guillaumerose 2021-05-19
> An alternative could be for the daemon to keep a long lasting ssh connection, so that we don't need to recreate it for every status call?

Yes, I tried this but the default Golang implementation doesn't have ssh keepalive and we have to do it by ourselves. 
Something like https://github.com/shazow/ssh-chat/pull/147/files#diff-b9630fed085ae953f0961a2b245323823c700efdbb43fe85b03bc229c1d8a5bcR83
Then we have to deal with reconnection, etc. Not really easy!

> Or would it be too slow anyway?

I think the second call to df is pretty fast. 

> But does it need the full status all the time? I don't think disk size is required for each of these polls, so there could be a 'fast status' command and a slower one?

Yes it can be a solution. The approach in this PR will also work if we change `crc status` to call the daemon.