@cfergeau 2021-05-19
`Get the cluster operator list via the http API and not ssh` makes the crc binary 30% to 50% bigger :-/ Goes from 38MB to 52MB